RWS class extracted from SystemConstants class and moved to rws.h
