+++ CMakeLists.txt
@@ -47,6 +47,7 @@ set(
     src/rws_state_machine_interface.cpp
     src/rws_subscription.cpp
     src/rws_websocket.cpp
+    src/rws.cpp
 )
 
 add_library(${PROJECT_NAME} ${SRC_FILES})
